// src/types/singer.d.ts
/** @format */

export declare type SingerInfo = {
	alias: string[];
	name: string;
	briefDesc: string;
	id: number;
	followed: boolean;
	picUrl: string;
	fullName: string;
};

export declare type SearchSingerItem = {
	accountId: number;
	albumSize: number;
	alia: string[];
	alias: string[];
	followed: boolean;
	id: number;
	picUrl: string;
	img1v1Url: string;
	mvSize: number;
	name: string;
	trans: string | null;
	// transNames: string[];
};